Initial commit of create-cs2007-wgem-combined-schema.py
This Python script will be created to merge the data from the CS2007 restored schema and the WGEM schema into a single related tables schema suitable for use with the Sweet mapping application.  The script was developed from the create-arcgis-online-data-model-20160728-01.py Python script developed to create a related tables model for Sweet Mapping for the Summer 2016 Pilot.